Add dry run

+++ boilerplate_gen/generator.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from boilerplate_gen.list import list_bp
 from boilerplate_gen.info import info_one
 
@@ -17,9 +18,11 @@ def generate(arguments, appconfig):
   if not os.path.isabs(arguments.path):
     arguments.path = os.path.abspath(arguments.path)
     
-    do_generate(arguments.boilerplate, arguments.path, appconfig)
+    do_generate(arguments.boilerplate, arguments, appconfig)
 
-def do_generate(boilerplate, path, appconfig):
+def do_generate(boilerplate, arguments, appconfig):
+  path = arguments.path
+  dry_run = arguments.dry_run
   print('Generating boilerplate for ' + boilerplate + ' on ' + path)
   print(" ")
   # Get config
@@ -30,7 +33,7 @@ def do_generate(boilerplate, path, appconfig):
   # Later: Validate all the depends_on section is installed
 
   # Prompt user for necessary info and set env vars
-  set_proper_env(bp_config)
+  set_proper_env(bp_config, dry_run)
 
   bp_dir = appconfig['core_dir'] + os.sep + boilerplate 
   before_copy_script = bp_dir + os.sep + bp_config.get('boilerplate', 'before_copy', fallback='before_copy.py')
@@ -47,43 +50,54 @@ def do_generate(boilerplate, path, appconfig):
 
   # Run the before copy script if it exists
   if os.path.isfile(before_copy_script) and os.access(before_copy_script, os.X_OK):
-    output = out(before_copy_script)
+    print("Running before copy script")
+    output = out(before_copy_script, dry_run)
     # Probably needs a flag later on
-    print("[OUTPUT FROM BEFORE_COPY_SCRIPT]: ", output)
   else:
     # @todo: if some flag is passed, stop execution here
     if os.path.isfile(before_copy_script) and not os.access(before_copy_script, os.X_OK):
-      print("before_copy script is not executable. Please contact boilerplate maintainer")
+      print("before_copy script is not executable. Please contact boilerplate maintainer. Exiting")
+      return
     else:
       print('Either before_copy script was not provided or we could not find it. Skipping')
 
   # Get the files to copy
   file_path = bp_dir + os.sep + bp_config.get('boilerplate', 'file_path', fallback='files')
 
+  # Try to create dir first
+  try:
+    if not dry_run:
+      os.mkdir(path)
+  except Exception:
+    pass
+
   # Copy them
-  copy_tree(file_path, path)
+  print("Copying the file tree")
+  if not dry_run:
+    copy_tree(file_path, path)
 
   # Run the after copy script if it exists
   if os.path.isfile(after_copy_script) and os.access(after_copy_script, os.X_OK):
-    output = out(after_copy_script)
-    # Probably needs a flag later on
-    print("[OUTPUT FROM AFTER_COPY]: ", output)
+    print("Running after copy script")
+    out(after_copy_script, dry_run)
   else:
     # @todo: if some flag is passed, stop execution here
     if os.path.isfile(after_copy_script) and not os.access(after_copy_script, os.X_OK):
-      print("after_copy script is not executable. Please contact boilerplate maintainer")
+      print("after_copy script is not executable. Please contact boilerplate maintainer. Exiting")
+      return
     else:
       print('Either after_copy script was not provided or we could not find it. Skipping')
 
 
-  # Change to the user specified path 
-  os.chdir(path)
+  # Change to the user specified path
+  if not dry_run:
+    os.chdir(path)
 
   # Perform installation, if script supplied
   if os.path.isfile(during_install_script) and os.access(during_install_script, os.X_OK):
-    output = out(during_install_script)
+    print("Running during install script")
+    output = out(during_install_script, dry_run)
     # Probably needs a flag later on
-    print("[OUTPUT FROM DURING_INSTALL]: ", output)
   else:
     # @todo: if some flag is passed, stop execution here
     if os.path.isfile(during_install_script) and not os.access(during_install_script, os.X_OK):
@@ -94,11 +108,18 @@ def do_generate(boilerplate, path, appconfig):
 
 
 
-def out(command):
-  result = run(command, stdout=PIPE, stderr=PIPE, universal_newlines=True, shell=True)
-  return result.stdout
+def out(command, dry_run = False):
+  if not dry_run:
+    result = run(command, universal_newlines=True, shell=True)
+  # return result.stdout
 
-def set_proper_env(config: configparser.ConfigParser):
+def check_deps(config: configparser.ConfigParser, dry_run):
+  if not config.has_section('depends_on'):
+    return
+  
+
+
+def set_proper_env(config: configparser.ConfigParser, dry_run):
   if not config.has_section('user_vars'):
     print("Did not find user_vars section on config, moving ahead")
     return
@@ -111,7 +132,8 @@ def set_proper_env(config: configparser.ConfigParser):
   prefix = 'PYTHON_BOILERPLATE_GEN_USERVAR_'
   for k, v in final_vars.items():
     print("Setting var:", prefix + k.upper(), "to ", v)
-    os.environ[prefix + k.upper()] = v
+    if (not dry_run):
+      os.environ[prefix + k.upper()] = v
   
 
 def get_value_from_user(k, v):
@@ -130,7 +152,7 @@ def get_required_value(k, v):
     return value.strip()
 
 def get_optional_value(k, v):
-  value = input("Enter value for" + k + "[optional]: ")
+  value = input("Enter value for " + k + " [optional]: ")
 
   if value.strip() == '':
     if len(v) > 1:
